Refactor player to use fewer threads
The previous form factor used N producer threads for N sequences. However
the synchronization wasn't straightforward and there's really no reason to
use so many threads. So to simplify, we just render each Midibox to a finite
reel of notes to play and then keep track of the playheads for each channel
inside of the player.